Added support for SVM using the libsvm-ruby-swig library.
No performance optimization yet.  Will do some crude performance tests first.

Tried to use the original libsvm-ruby by Rudi Cilibrar, but it cannot be
downloaded, nothing is even listening on port 80 on the server.  Also, it's
GPLed, which means that this library would have to be GPLed, and then anything
built with this library would have to be GPLed, too.  Do not want.